라우팅 카테고리에 맞게 분리 통합
