+++ package.json
@@ -16,6 +16,7 @@
     "@testing-library/user-event": "^13.2.1",
     "axios": "^0.26.0",
     "dotenv": "^16.0.0",
+    "qs": "^6.10.3",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-google-login": "^5.2.2",
