+++ src/App.js
@@ -3,7 +3,7 @@ import { Route, Switch } from "react-router-dom";
 import { Header, NotFound } from "./NavComponents";
 import { TendencyTest, MyInterests, Results, DimoDetail, DimoQNADetail, InlineDetail } from "./components";
 import { CreateArtWork, CreateDimo } from "./creates"
-import { Main, ArtWork, DimoQNA, DimoShared, MySpace, MyPosts, CreateProfile, CompleteProfile, KakaoRedirectHandler, LogOut, EditMySpace } from "./page";
+import { Main, ArtWork, Dimo, MySpace, MyPosts, CreateProfile, CompleteProfile, KakaoRedirectHandler, LogOut, EditMySpace } from "./page";
 
 function App() {
   return (
@@ -14,24 +14,22 @@ function App() {
             <Route exact path="/" component={Main} />
             <Route exact path="/Main" component={Main} />
             <Route exact path="/logout" component={LogOut} />
-            <Route exact path="/artwork" component={ArtWork} />
-            <Route exact path="/dimoqna" component={DimoQNA} />
-            <Route exact path="/dimoshared" component={DimoShared} />
-            <Route exact path="/myspace" component={MySpace} />
-            <Route exact path="/editmyspace" component={EditMySpace} />
             <Route exact path="/myposts" component={MyPosts} />
+            
+            <Route path={["/artwork", "/createart", "/detail"]} component={ArtWork} />
+
+            <Route path={['/dimoqna', '/dimoshared', "/dimodetail", "/dimoqnadetail", "/createdimo/qna", "/createdimo/shared"]} component={Dimo} />
+
+            <Route exact path={["/myspace", "/dpolio", "/mypage", "editmyspace"]} component={MySpace} />
+            
+            
             <Route exact path="/tendencytest" component={TendencyTest} />
             <Route exact path="/results" component={Results} />
             <Route exact path="/myinterests" component={MyInterests} />
 
-            <Route exact path="/createart" component={CreateArtWork} />
             <Route exact path="/createprofile" component={CreateProfile} />
             <Route exact path="/completeprofile" component={CompleteProfile} />
-            <Route exact path="/createdimo/qna" component={CreateDimo} />
-            <Route exact path="/createdimo/shared" component={CreateDimo} />
-            <Route exact path="/detail" component={InlineDetail} />
-            <Route exact path="/dimodetail" component={DimoDetail} />
-            <Route exact path="/dimoqnadetail" component={DimoQNADetail} />
+
             <Route path="/user/kakao/callback" component={KakaoRedirectHandler}/>
 
             <Route component={NotFound} />
