+++ src/NavComponents/index.js
@@ -1,8 +1,7 @@
 import Header from "./Header";
 import AdjBar from "./AdjBar";
-import DimoHeader from "./DimoHeader";
 import Tabs from "./Tabs";
 import MiniTabs from "./MiniTabs";
 import NotFound from "./NotFound";
 
-export { Header, AdjBar, DimoHeader, Tabs, MiniTabs, NotFound }
+export { Header, AdjBar, Tabs, MiniTabs, NotFound }
