+++ src/components/Sliders/DimoSlider.js
@@ -8,13 +8,10 @@ import { Image, Text, Icon } from "../../elements";
 // 리액트 슬라이더 중에 제일 많이들쓰는거
 // 구글링해서 커스텀 css 만들기!
 export const Slide = tw(Slider)`
-    w-[30.25rem] lg:w-[30.25rem] xl:w-[50.25rem] 2xl:w-[62.5rem]
-    ${(props) => (props.dimo ? `
-    bg-transparent mx-auto mt-10 col-start-2 
+    w-[30.25rem] lg:w-[30.25rem] xl:w-[50.25rem] 2xl:w-[78.5rem]
+    mx-auto mt-10 col-start-2 
     col-end-6 row-start-1 col-opacity-90 text-white
-    font-sanss2 text-md w-[30.25rem]` : "")};
-    ${(props) => (props.main ? `bg-transparent mx-auto
-    text-white overflow-hidden text-lg` : "")};
+    overflow-hidden text-lg
 `
 
 
@@ -22,16 +19,15 @@ const SS = tw.div`
     
 `
 const PrevBtn = tw.button`
-  z-10 text-white mt-36
+  z-10 text-black 
 `
 
 const NextBtn = tw.button`
-  z-10 text-white mt-36
+  z-10 text-black 
 `
 
-const MainSlider = (props)=>{
+const DimoSlider = (props)=>{
     const slider = React.useRef(null);
-    const { main, dimo } = props;
 
     const settings = {
         dots: false,  // 슬라이드 밑에 점 보이게
@@ -69,40 +65,11 @@ const MainSlider = (props)=>{
        ]
      }
 
-    
-    if(main) return (
-      <>
-      <div className="flex-row hidden md:flex">
-       <PrevBtn onClick={() => slider?.current?.slickPrev()}><Icon name="ArrowL" iconSize="48" /> </PrevBtn>
-    	<Slide {...settings} ref={slider}>
-         
-                <SS> 
-                    <Slides main />
-                </SS>
-                <SS>
-                    <Slides main />
-                </SS>
-                <SS>
-                    <Slides main />
-                </SS>
-                <SS>
-                    <Slides main />
-                </SS>
-                <SS>
-                    <Slides main />
-                </SS>
-               
-            </Slide>
-            <NextBtn onClick={() => slider?.current?.slickNext()}><Icon name="ArrowR" iconSize="48" /></NextBtn>
-            </div>  
-            </>
-            
-    );
-
    
-    if(dimo) return (
+   return (
       <>
-       <PrevBtn onClick={() => slider?.current?.slickPrev()}>◀</PrevBtn>
+        <div className="flex-row hidden md:flex">
+       <PrevBtn onClick={() => slider?.current?.slickPrev()}><Icon name="ArrowL" iconSize="48" /> </PrevBtn>
     	<Slide {...settings} ref={slider}>
          
                 <SS> 
@@ -126,13 +93,13 @@ const MainSlider = (props)=>{
                 </SS>
                 <SS>
                
-                <Slides dimo />
+                <Slides dimo  />
                 
                 </SS>
                
             </Slide>
-            <NextBtn onClick={() => slider?.current?.slickNext()}>▶</NextBtn> 
-            
+            <NextBtn onClick={() => slider?.current?.slickNext()}><Icon name="ArrowR" iconSize="48" /></NextBtn>
+            </div>  
             </>
             
     );
@@ -140,10 +107,5 @@ const MainSlider = (props)=>{
    
 }
 
-MainSlider.defaultProps = {
-  main: false,
-  dimo: false,
-  artwork: true,
-};
 
-export default MainSlider;
+export default DimoSlider;
