+++ src/components/index.js
@@ -9,25 +9,28 @@ import DimoWork from "./Posts/DimoWork";
 import MyArtWork from "./Posts/MyArtWork";
 
 import AllList from "./Lists/AllList";
-import DimoList from "./Lists/DimoList";
+import DimoList from "./Dimo/DimoList";
 
 import MainSlider from "./Sliders/MainSlider";
 import DetailSlider from "./Sliders/DetailSlider"
-
+import DimoSlider from "./Sliders/DimoSlider"
 import MySpaceUser from "./MySpace/MySpaceUser";
 import MyExp from "./MySpace/MyExp";
 import MyPortfolio from "./MySpace/MyPortfolio"
 import MySkill from "./MySpace/MySkill"
 import MyTimeLine from "./MySpace/MyTimeLine"
 import Dpolio from "./MySpace/Dpolio";
+import MyPage from "./MySpace/MyPage";
 
 import Results from "./Tendency/Results";
 import TendencyTest from "./Tendency/TendencyTest";
 import MyInterests from "./Tendency/MyInterests";
 
+import DimoCategory from "./Dimo/DimoCategory";
+import Categorys from "./Artwork/Categorys"
 
-export { ArtWorkDetail, DimoDetail, DimoQNADetail, InlineDetail, Comment, DetailSlider,
+export { ArtWorkDetail, DimoDetail, DimoQNADetail, InlineDetail, Comment, DetailSlider, DimoSlider, DimoCategory,
     ArtWork, DimoWork, MyArtWork, 
-    AllList, DimoList, 
+    AllList, DimoList, Categorys,
     MainSlider, MySpaceUser, MyExp, MyPortfolio, MySkill, MyTimeLine, 
-    Results, TendencyTest, MyInterests, Dpolio };
+    Results, TendencyTest, MyInterests, Dpolio, MyPage };
