+++ src/elements/Common/Card.js
@@ -17,7 +17,7 @@ m-2 rounded-lg justify-center flex flex-col items-center
 border-y-yellow-200 
 ${(props) => (props.size === "1" ? `w-c01 h-c01` : "")};
 ${(props) => (props.size === "2" ? `w-c02 h-c02` : "")};
-${(props) => (props.size === "3" ? `w-c03 h-c03 border-y-2` : "")};
+${(props) => (props.size === "3" ? `w-c03 h-c03 bg-dpurple-200 shadow-md` : "")};
 ${(props) => (props.size === "4" ? `w-c04 h-c04` : "")};
 ${(props) => (props.size === "5" ? `w-c05 h-c05` : "")};
 ${(props) => (props.size === "6" ? `w-c06 h-c06` : "")};
